int to long long

+++ codeC/headers/avl.h
@@ -3,21 +3,21 @@
 
 typedef struct _avl {
     int id;
-    int load;
-    int capacity;
+    long long load;
+    long long capacity;
     struct _avl* pLeft;
     struct _avl* pRight;
     int balance;
 } AVL;
 
-AVL* createAVL(int i, int c, int l);
+AVL* createAVL(int i, long long c, long long l);
 AVL* rotateLeft(AVL* pRoot);
 AVL* rotateRight(AVL* pRoot);
 AVL* doubleRotateLeft(AVL* pRoot);
 AVL* doubleRotateRight(AVL* pRoot);
 AVL* balanceAVL(AVL* pRoot);
 int searchAVL(AVL* pRoot, int i);
-AVL* insertAVL(AVL* pRoot, int i, int c, int l, int* h);
+AVL* insertAVL(AVL* pRoot, int i, long long c, long long l, int* h);
 void infix(AVL* pRoot);
 void freeAVL(AVL* pRoot);
